requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-WTF==1.2.1
email-validator==2.1.0.post1
pymongo==4.6.3
python-dotenv==1.0.1

# Text extraction
pdfminer.six==20231228
python-docx==1.1.0
textract==1.6.5

# NLP / AI
nltk==3.9.1
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.1
sentence-transformers==3.0.1

# Visualization & Reporting
matplotlib==3.8.4
reportlab==4.2.2